% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barnes2008.R
\docType{data}
\name{Barnes2008}
\alias{Barnes2008}
\title{Predator and prey body sizes}
\format{A data frame with 2 variables}
\usage{
data(Barnes2008)
}
\description{
A dataset containing only the predator and prey lengths from Barnes et al 2008.
}
\details{
\itemize{
  \item standardised_predator_length: Predator length
  \item si_prey_length: Prey length  
  }
}
\keyword{datasets}

