% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bartomeus2008.R
\docType{data}
\name{Bartomeus2008}
\alias{Bartomeus2008}
\title{Plant and pollinator sizes}
\format{A data frame with 6 variables}
\usage{
data(Bartomeus2008)
}
\description{
A dataset containing plant and pollinator lengths from Bartomeus et al 2008.
}
\details{
\itemize{
  \item plant: Plant latin name
  \item nectar_holder_depth_mm: Plant nectar holder depth in mm 
  \item pollinator: Pollinator latin name
  \item IT_mm: Pollinator Intertegular span in mm
  \item family: Pollinator family  
  \item cover: Plant cover from independent transects  
  }
}
\keyword{datasets}

