% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deraison2014.R
\docType{data}
\name{Deraison2014}
\alias{Deraison2014}
\title{Plant and grasshopper traits}
\format{A data frame with 16 variables}
\usage{
data(Deraison2014)
}
\description{
A dataset containing plant and grasshopper interactions and traits from Deraison et al 2014.
}
\details{
\itemize{
  \item Plant.species: Plant ID
  \item Grasshopper.species: Grasshopper ID
  \item Herbivory: % of plant consumed
  \item G.Body.size.mm: Grasshopper body size in mm
  \item G.Incisive.strength: Grasshopper incisive strength
  \item G.Carbon.nitrogen.ratio: Plant C/N ratio 
  \item G.Species.name: Grasshopper latin name
  \item P.Leaf.dry.matter.content: Plant leaf dry matter content  
  \item P.Leaf.nitrogen.content.dry.mass:  
  \item P.Leaf.carbon.content.dry.mass:  
  \item P.Leaf.carbon.nitrogen.ratio:   
  \item P.Leaf.thickness.mm:   
  \item P.Leaf.area.cm:  
  \item P.Leaf.thickness.mm:   
  \item P.Perimeter.leaf.lenght.ratio: 
  \item P.Species.name: Plant latin name  
  }
}
\keyword{datasets}

