% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/FitModel.R
\name{fit_it}
\alias{fit_it}
\title{Fit a model using SA}
\usage{
fit_it(model, Tlevel1, Tlevel2, mean_Tlevel1, sd_Tlevel1)
}
\arguments{
\item{model}{}

\item{Tlevel1}{}

\item{Tlevel2}{}

\item{mean_Tlevel1}{}

\item{sd_Tlevel1}{}
}
\value{
x
}
\description{
.
}
\examples{
x
}
\author{
Dominique Gravel and Ignasi Bartomeus
}
\references{
Williams et al 2010?
}

