% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Model_GenSA.R
\name{integrated_model}
\alias{integrated_model}
\title{Bayesian model of species interactions based on traits, using observation of interactions only (presence)}
\usage{
integrated_model(pars, Tlevel1, Tlevel2, mean_Tlevel1, sd_Tlevel1)
}
\arguments{
\item{pars}{}

\item{Tlevel1}{}

\item{Tlevel2}{}

\item{mean_Tlevel1}{}

\item{sd_Tlevel1}{}
}
\value{
x
}
\description{
.
}
\examples{
x
}
\author{
Dominique Gravel
}
\references{
Williams et al 2010?
}

