% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{integrated_model}
\alias{models}
\title{Bayesian models for characterizing trait-based species interactions.}
\usage{
integrated_model(pars, Tlevel1, Tlevel2, mean_Tlevel1, sd_Tlevel1)
}
\description{
This models are described in Bartomeus at al. (2015 Functional Ecology) and are based
on Williams et al. 2000 models. They are intended to be used with fit_it function. The integrated model
considers both neutral and niche constrains, while the neutral and niche models only consider its respective
components.
}
\author{
Dominique Gravel
}
\references{
Williams, R.J., Anandanadesan, A. & Purves, D. (2010) The probabilistic niche model reveals the niche structure and role of body size in a complex food web. PloS one, 5, e12092.
Williams, R.J. & Martinez, N.D. (2000) Simple rules yield complex food webs. Nature, 404, 180–183.#'
}

