% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plot.pred.R
\name{plot.pred}
\alias{plot.pred}
\title{Plot model output}
\usage{
\method{plot}{pred}(pars, Tlevel1, Tlevel2, xlab = "Trait level 2",
  ylab = "Trait level 1")
}
\arguments{
\item{pars}{}

\item{Tlevel1}{}

\item{Tlevel2}{}

\item{xlab}{}

\item{ylab}{}
}
\value{
x
}
\description{
.
}
\examples{
x
}
\author{
Dominique Gravel and Ignasi Bartomeus
}
\references{
Williams et al 2010?
}

