% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tylinakis2008.R
\docType{data}
\name{Tylianakis2008}
\alias{Tylianakis2008}
\title{Host and parasitoid traits}
\format{A data frame with 5 variables}
\usage{
data(Tylianakis2008)
}
\description{
A dataset containing host and parasitoid interactions and traits from Tylianakis et al 2008.
}
\details{
\itemize{
  \item parasite: Parasite ID
  \item host: Host ID
  \item freq: number of interactions recorded
  \item host_body_length: Host body length in mm
  \item parasite_body_length: Parasite body length in mm 
  }
}
\keyword{datasets}

