% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/useful_functions.R
\name{Useful functions}
\alias{Useful functions}
\alias{weighted.mean}
\title{weighted mean and sd}
\usage{
weighted.mean(x, w)
}
\arguments{
\item{x}{.}

\item{w}{}
}
\value{
x
}
\description{
just weighted mean and sd.
}
\examples{
x
}

